:recycle: Reorganize Dynamic AssetType functions
