+++ transactions/createAssetType.go
@@ -2,14 +2,14 @@ package transactions
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/goledgerdev/cc-tools/assets"
 	"github.com/goledgerdev/cc-tools/errors"
 	sw "github.com/goledgerdev/cc-tools/stubwrapper"
 )
+
+// ? Hanlde Validate function fields?
 
 // CreateAssetType is the transaction which creates a dynamic Asset Type
 var CreateAssetType = Transaction{
@@ -34,7 +34,7 @@ var CreateAssetType = Transaction{
 		for _, assetType := range assetTypes {
 			assetTypeMap := assetType.(map[string]interface{})
 
-			newAssetType, err := BuildAssetType(assetTypeMap)
+			newAssetType, err := buildAssetType(assetTypeMap)
 			if err != nil {
 				return nil, errors.WrapError(err, "failed to build asset type")
 			}
@@ -57,8 +57,8 @@ var CreateAssetType = Transaction{
 	},
 }
 
-func BuildAssetType(typeMap map[string]interface{}) (assets.AssetType, errors.ICCError) {
-	// *********** Build Props Array ***********
+func buildAssetType(typeMap map[string]interface{}) (assets.AssetType, errors.ICCError) {
+	// Build Props Array
 	propsArr, ok := typeMap["props"].([]interface{})
 	if !ok {
 		return assets.AssetType{}, errors.NewCCError("invalid props array", http.StatusBadRequest)
@@ -66,28 +66,27 @@ func BuildAssetType(typeMap map[string]interface{}) (assets.AssetType, errors.IC
 	props := make([]assets.AssetProp, len(propsArr))
 	for i, prop := range propsArr {
 		propMap := prop.(map[string]interface{})
-		assetProp, err := BuildAssetProp(propMap)
+		assetProp, err := assets.BuildAssetProp(propMap)
 		if err != nil {
 			return assets.AssetType{}, errors.WrapError(err, "failed to build asset prop")
 		}
 		props[i] = assetProp
 	}
 
-	// *********** Check Type Values ***********
 	// Tag
-	tagValue, err := CheckValue(typeMap["tag"], true, "string", "tag")
+	tagValue, err := assets.CheckValue(typeMap["tag"], true, "string", "tag")
 	if err != nil {
 		return assets.AssetType{}, errors.WrapError(err, "invalid tag value")
 	}
 
 	// Label
-	labelValue, err := CheckValue(typeMap["label"], true, "string", "label")
+	labelValue, err := assets.CheckValue(typeMap["label"], true, "string", "label")
 	if err != nil {
 		return assets.AssetType{}, errors.WrapError(err, "invalid label value")
 	}
 
 	// Description
-	descriptionValue, err := CheckValue(typeMap["description"], false, "string", "description")
+	descriptionValue, err := assets.CheckValue(typeMap["description"], false, "string", "description")
 	if err != nil {
 		return assets.AssetType{}, errors.WrapError(err, "invalid description value")
 	}
@@ -100,12 +99,12 @@ func BuildAssetType(typeMap map[string]interface{}) (assets.AssetType, errors.IC
 		// Validate
 	}
 
-	// *********** Build Readers Array ***********
+	// Readers
 	readers := make([]string, 0)
 	readersArr, ok := typeMap["readers"].([]interface{})
 	if ok {
 		for _, reader := range readersArr {
-			readerValue, err := CheckValue(reader, false, "string", "reader")
+			readerValue, err := assets.CheckValue(reader, false, "string", "reader")
 			if err != nil {
 				return assets.AssetType{}, errors.WrapError(err, "invalid reader value")
 			}
@@ -113,154 +112,9 @@ func BuildAssetType(typeMap map[string]interface{}) (assets.AssetType, errors.IC
 			readers = append(readers, readerValue.(string))
 		}
 	}
-
 	if len(readers) > 0 {
 		assetType.Readers = readers
 	}
 
 	return assetType, nil
-}
-
-func BuildAssetProp(propMap map[string]interface{}) (assets.AssetProp, errors.ICCError) {
-	// *********** Check Prop Values ***********
-	// Tag
-	tagValue, err := CheckValue(propMap["tag"], true, "string", "tag")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid tag value")
-	}
-
-	// Label
-	labelValue, err := CheckValue(propMap["label"], true, "string", "label")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid label value")
-	}
-
-	// Description
-	descriptionValue, err := CheckValue(propMap["description"], false, "string", "description")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid description value")
-	}
-
-	// Required
-	requiredValue, err := CheckValue(propMap["required"], false, "boolean", "required")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid required value")
-	}
-
-	// IsKey
-	isKeyValue, err := CheckValue(propMap["isKey"], false, "boolean", "isKey")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid isKey value")
-	}
-
-	// ReadOnly
-	readOnlyValue, err := CheckValue(propMap["readOnly"], false, "boolean", "readOnly")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid readOnly value")
-	}
-
-	// DataType
-	dataTypeValue, err := CheckValue(propMap["dataType"], true, "string", "dataType")
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "invalid dataType value")
-	}
-
-	err = CheckDataType(dataTypeValue.(string))
-	if err != nil {
-		return assets.AssetProp{}, errors.WrapError(err, "failed checking data type")
-	}
-
-	assetProp := assets.AssetProp{
-		Tag:         tagValue.(string),
-		Label:       labelValue.(string),
-		Description: descriptionValue.(string),
-		Required:    requiredValue.(bool),
-		IsKey:       isKeyValue.(bool),
-		ReadOnly:    readOnlyValue.(bool),
-		DataType:    dataTypeValue.(string),
-		// Validate
-	}
-
-	// *********** Build Writers Array ***********
-	writers := make([]string, 0)
-	writersArr, ok := propMap["writers"].([]interface{})
-	if ok {
-		for _, writer := range writersArr {
-			writerValue, err := CheckValue(writer, false, "string", "writer")
-			if err != nil {
-				return assets.AssetProp{}, errors.WrapError(err, "invalid writer value")
-			}
-
-			writers = append(writers, writerValue.(string))
-		}
-	}
-	if len(writers) > 0 {
-		assetProp.Writers = writers
-	}
-
-	// ********* Validate Default Value *********
-	if propMap["defaultValue"] != nil {
-		// TODO: Reorganize utils functions and return ValidateProp to protected (validateProp)
-		defaultValue, err := assets.ValidateProp(propMap["defaultValue"], assetProp)
-		if err != nil {
-			return assets.AssetProp{}, errors.WrapError(err, "invalid Default Value")
-		}
-
-		assetProp.DefaultValue = defaultValue
-	}
-
-	return assetProp, nil
-}
-
-func CheckDataType(dataType string) errors.ICCError {
-	trimDataType := strings.TrimPrefix(dataType, "[]")
-
-	if strings.HasPrefix(trimDataType, "->") {
-		trimDataType = strings.TrimPrefix(trimDataType, "->")
-
-		assetType := assets.FetchAssetType(trimDataType)
-		if assetType == nil {
-			return errors.NewCCError(fmt.Sprintf("invalid dataType value %s", dataType), http.StatusBadRequest)
-		}
-	} else {
-		dataTypeObj := assets.FetchDataType(trimDataType)
-		if dataTypeObj == nil {
-			return errors.NewCCError(fmt.Sprintf("invalid dataType value %s", dataType), http.StatusBadRequest)
-		}
-	}
-
-	return nil
-}
-
-func CheckValue(value interface{}, required bool, expectedType, fieldName string) (interface{}, errors.ICCError) {
-	if value == nil {
-		if required {
-			return nil, errors.NewCCError(fmt.Sprintf("required value %s missing", fieldName), http.StatusBadRequest)
-		}
-		switch expectedType {
-		case "string":
-			return "", nil
-		case "number":
-			return 0, nil
-		case "boolean":
-			return false, nil
-		}
-	}
-
-	switch expectedType {
-	case "string":
-		if _, ok := value.(string); !ok {
-			return nil, errors.NewCCError(fmt.Sprintf("value %s is not a string", fieldName), http.StatusBadRequest)
-		}
-	case "number":
-		if _, ok := value.(float64); !ok {
-			return nil, errors.NewCCError(fmt.Sprintf("value %s is not a number", fieldName), http.StatusBadRequest)
-		}
-	case "boolean":
-		if _, ok := value.(bool); !ok {
-			return nil, errors.NewCCError(fmt.Sprintf("value %s is not a boolean", fieldName), http.StatusBadRequest)
-		}
-	}
-
-	return value, nil
 }
